Update launch.json: Use integrated terminal for console.
This commit updates the launch.json file by adding the "console": "integratedTerminal" configuration. This change allows the integrated terminal to be used as the console when launching the program.

Update schema.prisma: Add comments and fields for guildusers model.

This commit adds comments and fields for the guildusers model in the schema.prisma file. The model now has comments in the database and requires additional setup for migrations. The new fields include user_id and nickname.

Rename Database.go to User.go and Guild.go to Database/Guild.go.

This commit renames the Database.go file to User.go and the Guild.go file to Database/Guild.go in the src/Database directory. The purpose of this change is to improve code organization and maintain consistency in file naming conventions.

Rename Guild/Database.go to Database/Guild.go.

This commit renames the Guild/Database.go file to Database/Guild.go. The purpose of this change is to improve code organization and maintain consistency in file naming conventions.

Delete Guild/Search.go.

This commit deletes the Guild/Search.go file. The file is no longer needed and can be safely removed.

Update Database.go: Remove verbose log statements.

This commit removes the verbose log statements in the InsertUser function of the Database.go file. The purpose of this change is to reduce